election.c
Instructor's description of program:

Write a program to do the following:
a. Display the table with appropriate labels for the rows and columns.
b. Compute and display the total number of votes received by each
candidate and the percentage of the total votes cast.
c. If any one candidate received over 50 percent of the votes, the
program should display a message declaring that the candidate the
winner.
d. If no candidate received 50 percent of the votes, the program should
display a message declaring a runoof between the two candidates
receiving the highest number of votes; the two candidates should be
identified by their letter names.
e. Run the program once with the data shown and once with candidate C
receiving only 108 votes in Precinct 4.”
Additional specifications: The data will be stored in a file that will
be passed as the only command line parameter. The format of the file
will match that of the table above, except that the labels will not be
in the file. Instead of presenting the candidates in alphabetical order,
they should be presented from the largest total to smallest. You must
write a sort() function that sorts the election matrix based on the
total votes of each candidate. Your main will call four functions:
read_file(), sort(), show_table, and declare_winner(). Your main() may
have nothing in it but variable declarations, the our function calls,
and the return statement. There will always be five precincts and four
candidates. Your output should match mine exactly.

Input files are election 1 2 3.txt