maze.c
Finds a correct path out  of a input maze file (where O denotes an open
path and X denotes a wall). Uses recursion to find the path out and
outputs the coordinates to take to reach the exit.

Maze 1 2 3 4.txt is the input maze text files.

I don't know what happened to my program (it's missing) so I'll be using
my instructor's file as I had the same file output and implementation as
him